new CultureInfo("nl-BE") bij elke ToString("C") gezet om een €-teken te krijgen

+++ src/KairosWeb_Groep6/Controllers/Baten/ExtraProductiviteitController.cs
@@ -39,7 +39,7 @@ namespace KairosWeb_Groep6.Controllers.Baten
                     {
                         Type = model.Type,
                         Soort = model.Soort,
-                        Bedrag = model.Bedrag
+                        Bedrag = (double) model.Bedrag
                     };
 
                     analyse.ExtraProductiviteit = baat;
